캐릭터 이동방식을 DoTween으로 변경, 플레이어 이동속도 변경 (int -> float)

+++ Assets/Script/Manager/FileManager.cs
@@ -256,7 +256,7 @@ public class FileManager
             }
             playerItem.PlayerMovingType = (MovingType)int.Parse(movingType);
             playerItem.Hp = int.Parse(hp);
-            playerItem.MoveSpeed = int.Parse(moveSpeed);
+            playerItem.MoveSpeed = float.Parse(moveSpeed);
             playerItem.MoveRange = int.Parse(moveRange);
             playerItem.AttackRange = int.Parse(attackRange);
             playerItems.Add(playerItem);
